Php7 (#63)
* implement php7 support, no backward compatibility

* fix memory leaks

* fix ProtobufMessage __construct method

* fix leak when opcache is enabled

* fix segfault, cover this case with test

* int type mismatch fixed

* fix internal int64 value representation for x86 systems

* add test for int64

* fix test

* fix test tests/parse_repeated_obj.phpt

* fix old macros usage

* update requirements in README

* Test parse "false" value after encode-decode

* fix bool value setting/serializing

* fix php 7.1 compatibility

* looping debug fixed

* fix infinite loop and add test for that

+++ reader.c
@@ -37,7 +37,7 @@ int reader_read_double(reader_t *reader, double *value)
 	return -1;
 }
 
-int reader_read_fixed32(reader_t *reader, long *value)
+int reader_read_fixed32(reader_t *reader, int32_t *value)
 {
 	fixed32_t val;
 
@@ -52,7 +52,7 @@ int reader_read_fixed32(reader_t *reader, long *value)
 	return -1;
 }
 
-int reader_read_fixed64(reader_t *reader, long *value)
+int reader_read_fixed64(reader_t *reader, int64_t *value)
 {
 	fixed64_t val;
 
@@ -82,7 +82,7 @@ int reader_read_float(reader_t *reader, double *value)
 	return -1;
 }
 
-int reader_read_int(reader_t *reader, long *value)
+int reader_read_int(reader_t *reader, int64_t *value)
 {
 	uint64_t val;
 
@@ -94,7 +94,7 @@ int reader_read_int(reader_t *reader, long *value)
 	return 0;
 }
 
-int reader_read_signed_int(reader_t *reader, long *value)
+int reader_read_signed_int(reader_t *reader, int64_t *value)
 {
 	uint64_t val;
 
@@ -132,7 +132,7 @@ int reader_read_string(reader_t *reader, char **string, int *len)
 	return 0;
 }
 
-int reader_read_tag(reader_t *reader, uint32_t *field_number, uint8_t *wire_type)
+int reader_read_tag(reader_t *reader, uint64_t *field_number, uint8_t *wire_type)
 {
 	uint64_t key;
